+++ src/main/kotlin/cloud/glitchdev/rfu/config/RFUSettings.kt
@@ -3,6 +3,7 @@ package cloud.glitchdev.rfu.config
 import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.config.categories.BackendSettings
 import cloud.glitchdev.rfu.config.categories.DevSettings
+import cloud.glitchdev.rfu.config.categories.DropsSettings
 import cloud.glitchdev.rfu.config.categories.GeneralFishing
 import cloud.glitchdev.rfu.config.categories.HotSpotSettings
 import cloud.glitchdev.rfu.config.categories.InkFishing
@@ -125,11 +126,12 @@ object RFUSettings : ConfigKt("rfu/settings") {
         }
 
         category(GeneralFishing)
-        category(RareScSettings)
         category(LavaFishing)
         category(HotSpotSettings)
         category(InkFishing)
         category(JerryFishing)
+        category(RareScSettings)
+        category(DropsSettings)
         category(OtherSettings)
         category(BackendSettings)
         category(DevSettings)