+++ src/main/kotlin/cloud/glitchdev/rfu/feature/drops/RareDropTitleAlert.kt
@@ -1,9 +1,9 @@
 package cloud.glitchdev.rfu.feature.drops
-
-import cloud.glitchdev.rfu.config.categories.GeneralFishing
+import cloud.glitchdev.rfu.config.categories.DropsSettings
 import cloud.glitchdev.rfu.events.managers.DropEvents
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
+import cloud.glitchdev.rfu.constants.Rarity
 import cloud.glitchdev.rfu.data.drops.DropManager
 import cloud.glitchdev.rfu.data.drops.DropRecord
 import cloud.glitchdev.rfu.utils.Title
@@ -14,13 +14,13 @@ import cloud.glitchdev.rfu.utils.dsl.toReadableString
 object RareDropTitleAlert : Feature {
     override fun onInitialize() {
         DropEvents.registerRareDropEvent { rareDrop, magicFind ->
-            if (rareDrop !in GeneralFishing.rareDrops || !GeneralFishing.rareDropTitleAlert) return@registerRareDropEvent true
+            if (rareDrop !in DropsSettings.rareDrops || !DropsSettings.rareDropTitleAlert) return@registerRareDropEvent true
             showTitleAlert(rareDrop.dropName, rareDrop.rarity.color.code, DropManager.dropHistory.getOrAdd(rareDrop).history, magicFind)
             return@registerRareDropEvent true
         }
 
         DropEvents.registerDyeDropEvent { dyeDrop, magicFind ->
-            if (dyeDrop !in GeneralFishing.dyeDrops || !GeneralFishing.rareDropTitleAlert) return@registerDyeDropEvent
+            if (dyeDrop !in DropsSettings.dyeDrops || !DropsSettings.rareDropTitleAlert) return@registerDyeDropEvent
             showTitleAlert(dyeDrop.dyeName, dyeDrop.rarity.color.code, DropManager.dropHistory.getOrAdd(dyeDrop).history, magicFind)
         }
 
@@ -36,7 +36,7 @@ object RareDropTitleAlert : Feature {
             "First Drop"
         }
 
-        val title = GeneralFishing.rareDropTitleFormat
+        val title = DropsSettings.rareDropTitleFormat
             .replace("{drop}", dropName)
             .replace("{dropcolor}", dropColor)
             .replace("{magic_find}", magicFind?.toString() ?: "0")
@@ -44,7 +44,7 @@ object RareDropTitleAlert : Feature {
             .replace("{time}", timeSinceLast)
             .replace("{total}", history.size.toString())
             .toMcCodes()
-        val subtitle = GeneralFishing.rareDropSubtitleFormat
+        val subtitle = DropsSettings.rareDropSubtitleFormat
             .replace("{drop}", dropName)
             .replace("{dropcolor}", dropColor)
             .replace("{magic_find}", magicFind?.toString() ?: "0")
@@ -52,6 +52,7 @@ object RareDropTitleAlert : Feature {
             .replace("{time}", timeSinceLast)
             .replace("{total}", history.size.toString())
             .toMcCodes()
+
 
         Title.showTitle(title, subtitle, fadeIn = 5, duration = 40, fadeOut = 5)
     }