+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changes:
 - Removed the 30 minute to expire message in the party finder creation area.
 - Made Reindrake not display the lootshare range
+- Reorganized settings, added rare drops category
 
 ### Fixes:
 - Made water hotspot particles even more strict for detections (When will this end?)